Adds dalli < 3 to the contrib-old appraisal
In addition, this adds another build line in the Rakefile to build dalli
against the old contrib. This change makes sure that the current
contrib uses dalli >= 3.0.0 as well.

To build the gemfiles within the docker container, the compose file
needed to declare the platform architecture so that it would work on M1
based apple computers.