Replace JRuby 9.2.0.0 in CI with JRuby 9.2.8.0
In #1409 we added JRuby 9.2.0.0 explicitly to CI, because there was a
behavior change between <= 9.2.8.0 and >= 9.2.9.0 and we wanted to
support customers running on either side of the behavior change.

Keeping JRuby 9.2.0.0 working is non-trivial work, since it's missing
hundreds of fixes and changes from the 9.2 branch. We already had to
change things to acommodate it in #1434, and while investigating
https://github.com/msgpack/msgpack-ruby/issues/239 I found YET another
issue where 9.2.0.0 is broken but later releases are fine.

Thus, since the only place in our test suite where we actually test for
different behavior is related to #1409, let's replace 9.2.0.0 in our
test suite with 9.2.8.0.

Hopefully 9.2.8.0 is in a bit of a better shape, and we'll avoid
needing to chase down any more 9.2.0.0-specific bugs.