RFC: Remove deprecated ability to provide `c.tracer.transport_options` as a hash, support proc only
Up until now, we supported two ways of passing custom configurations
to `transport_options`: either a hash, or as a proc.

The hash variant is not (that I could find) documented anywhere, and
is a subset of the proc variant, so I propose that we remove it in
time for 1.0.0.

This allows us to remove a lot of extra complexity around configuring
our transports, while keeping the same flexibility in place (the proc
variant).