Temporarily rename Datadog::Profiling::Flush to OldFlush
As part of the ongoing refactoring work to report profiling data using
libddprof, the `Flush` class is going to need to change its shape.

To enable incremental development of the new code side-by-side with
the old code, let's rename the existing class to `OldFlush` so that
the new code can take over its name.