Enable running profiler specs on macOS
Profiler is currently not supported on macOS due to missing libddprof
binaries for it, although if you locally compile libddprof it actually
works.

Thus, I've added a `DD_PROFILING_MACOS_TESTING` environment variable
that can be used to allow testing on macOS, but is disabled by default.

I also added a check that the profiler is properly available, when
we expect it to be available. Without this check, specs that run with
a profiler that is not available would fail later with confusing error
messages.