Record failure to compile profiling extension and present info when trying to load profiler
There are many reasons (see `native_extension_helpers.rb`) why we may
not be able to compile the profiling native extension.

During gem installation, we check them, and instead of failing to
install dd-trace-rb, we print a nice banner explaining what went
wrong.

Unfortunately, the nice banner is not shown during `gem install`
or `bundle install` by default. Instead, customers would run into
a cryptic failure regarding the native extension not being able
to be loaded.

To solve this, whenever we skip compilation, we record that info
on a file (`skipped_reason.txt`). Then, at runtime, we check if that
file exists, and if it does, we use that as a nice explanation of
what went wrong.

It's somewhat "icky" to generate files during installation and put
them in our source tree, but I couldn't think of a better approach.