Drop profiler support for macOS
Previously, the profiler would start on macOS, but only get wall time
and be missing CPU time data.

If you're a customer reading this, and this impacts you, please get
in touch!

We need to do this because with the move to `libddprof` we can't
support macOS until we also have `libddprof` binaries for macOS.

In practice, the profiler does work on macOS if one manually
compiles and packages a `libddprof` version for macOS. So I've added
an override that is expected to be used only during development.