Move ENVIRONMENT_AGENT_SETTINGS to class/specs that need it
The `ENVIRONMENT_AGENT_SETTINGS` constant was introduced to support
a few classes/specs that needed an `agent_settings` object  to fall
back to.

Since there are so little users of this constant (which is good),
let's remove it from the `AgentSettingsResolver` class and instead
move it to the class/specs that make use of it.

This way, the weight of legacy/weird behavior is on the classes that
need it, rather than on the `AgentSettingsResolver` class.