Print warning when transport_configuration_proc is used but do not raise error
In https://github.com/DataDog/dd-trace-rb/pull/1932 we did the
groundwork required to support parsing the unix domain socket
configuration from a c.tracing.transport_options configuration.

Thus, we can now support unix domain sockets configured by the user
without any other change (assuming that PR is merged before this
commit is).

The `transport_configuration_proc` can still contain weird transport
configurations, but if that's the case, we just ignore them and
print a warning. This is because we don't support every setting
a user may provide in that proc.

Worst case, for customers with a really custom config, is that
the profiler won't be able to report data, and the customer will need
to reach out to support for help updating it. (But we won't raise
an exception or stop the profiler from starting).