Add benchmark for `HttpTransport` class
This benchmark can be used to evaluate the performance and memory
correctness of the `HttpTransport` class.

In particular, it was used to fix a memory leak (and to validate
that we don't have others), and was used to validate the change
where `HttpTransport` creates a new exporter for every report,
rather than reusing exporters.