Fix breakage in CI integration app on Ruby 2.3
For some reason on Ruby 2.3 we're failing to compile google-protobuf
3.20.0.rc.2 . Since that version is not stable anyway, I've just added
it to our list of excluded versions.

Here is one such failure:
https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/5727/workflows/876c380c-091d-46b2-9462-d424fa24b413/jobs/212239

If the issue reappears with the "final" release of 3.20.0, we can
always revisit and consider other options.

(I'm hoping that by then we'll perhaps have dropped the need to have
google-protobuf for profiler, so I think it's not worth investing
a lot of effort on this.)