Resist missing ruleset file
Adjusts the behaviour from raising an exception to outputting an error
and allow the app to continue running. This also generalises the error
management beyond a ruleset file being missing and also takes into
account various invalid ruleset cases or libddwaf unavailability.

In these cases, in addition to the error, a warning is produced that
alerts the user that AppSec is subsequently disabled.