Import `rb_profile_frames` from Ruby sources
The `rb_profile_frames` method can be used to get stack traces for
a profiler, but it has a number of limitations and caveats which we'll
need to bridge, and thus we need to ship our own copy of it.

This is OK since upstream Ruby allows the source code to be used under
the terms of the BSD-2-Clause license, which I include in the file
with the copied snippets. I'll also run this through the internal
opensource group, in case they have any notes/concerns.