Tweak webrick logging in tests to expose issues
Previously we were sending all webrick logs to a `StringIO` that was
never checked, which meant that if something was wrong inside webrick,
we would not know about it (since webrick runs on a background thread).

Instead, I've configured webrick to log to stderr, and raised the log
level so only important issues get printed.