Add debug log message during profile encoding
This is similar to the one we had in
`Datadog::Profiling::Encoding::Profile::Protobuf`, but for the new
libddprof-based `StackRecorder`.

Unfortunately we don't get as many debug stats with libddprof as we
had with the previous code; at some point we may want to take a stab
at exposing that info in libddprof.