[PROF-5473] Release global VM lock during profile serialization
Releasing the global VM lock during serialization allows the VM to
continue executing Ruby code in parallel with libddprof serialization.

This reduces the latency impact of serialization: even if it takes
a few hundred milliseconds, as long as there are enough CPUs on the
machine, the Ruby app will be able to continue its work undisturbed.