Implement `ddtrace_thread_list` to replace the inaccessible `rb_thread_list` (incomplete)
Since `rb_thread_list` is not exposed by Ruby, here's my first attempt
at implementing one, inspired by both `rb_thread_list` (Ruby 2) and
`rb_ractor_thread_list` (Ruby 3).

Thus far it seems to work on all 3.x Rubies, as well as on
2.7, 2.6, 2.4, 2.3 and 2.2.

Ruby 2.5 is getting into an infinite loop (??? -- to be investigated)
and Ruby 2.1 of course is different from the later 2.x series and
of course we'll need to do something special for it. Sigh.