Remove appraisal gemfiles from tree, leave only lockfiles
We started checking in the `gemfiles/` folder in #1470, with the goal
of making our CI pipeline more stable -- e.g. so that PRs didn't
pick up new upstream releases that made them break for unrelated
reasons.

A big downside is that whenever we make even the tiniest of changes
in the `Appraisals` file, here comes huge diff of the same change
being applied dozens of times to the multiple gemfile variants.

This got me thinking: the part that guarantees CI stability is
having the `gemfile.lock` files checked in.
Could we get by without having the `.gemfile` files checked in, and
always regenerating them when needed?

Turns out we can! So here's a PR for it.

Upsides:
* A lot less files in the tree, and a lot less diff noise when
  adding new Ruby releases, gems, etc.

Downsides:
* You need to run `appraisal generate` locally to regenerate the
  `.gemfile`s. (This was already the case before #1470)
 This takes less than a second on my machine.