Bootstrap per-thread context and hashmap to store it
There will be a few things that we'll want to track for each
thread (such as "when did we last sample it").

This is similar to what we already did for the `OldStack`
collector, where we store these things as instance variables in
the `Thread` object.

Currently the struct is actually empty; I'll start adding
things there soon.

A big FIXME at this point is that we never remove dead
threads from the context.