Use `rb_memerror()` when handing out of memory issues
As pointed out by Marco in
<https://github.com/DataDog/dd-trace-rb/pull/2019#discussion_r874258181>
it's possible that when memory allocation fails, we can be in such
a state that casually raising an exception may not work at all.

Instead, let's use a function that Ruby exposes for this exact
situation -- `rb_memerror()` -- which is built to withstand such
memory issues.