Fix wrong class used when calling TypedData_Wrap_Struct
The reason we receive `klass` as an argument to the function registed
with `rb_define_alloc_func` is so that the correct class can get
created, even when someone subclasses us.

Because we were ignoring that argument and always using
`stack_recorder_class`, we broke any attempts to subclass our class.
Interestingly, Ruby actually detected and complained about this:

```ruby
[4] pry(main)> class Foo < Datadog::Profiling::StackRecorder; end
=> nil
[5] pry(main)> Foo.new
TypeError: wrong instance allocation
```

NOT that we ever expect `StackRecorder` to be subclassed, but
it may come handy for some future experiementation or specs AND
it's helpful to have this correct because in the future we may
copy-paste it as an example of using `TypeddData_Wrap_Struct`,
and I don't like leaving a wrong example in.