Move ready signal closer to sleep
What we're doing in these tests is inherently a race: we want the
thread to be stopped at the `sleep`, but have no way to atomically
signal the `ready_queue` and go to `sleep`, and thus we kinda
pray that Ruby just won't schedule us away between both operations.

This seems to work so far, but to avoid creating any flaky tests,
let's try to signal readyness as close as possible to calling `sleep`.

(Note that if this ever gets us into trouble and ends up creating
test flakiness, maybe we could implement a native method that
signals the queue AND calls sleep and ensures the Global VM
Lock is not released.)