Update libddwaf to 1.3.0.2.0
This version bundles two dedicated builds one for each of glibc and
musl. This appears to resolve the segfault happening randomly on Ruby
2.6.

Also, while the java gem is unpublished yet on rubygems, this
version adds preliminary JRuby support, so manually building and
installing the corresponding gem allows for testing AppSec on JRuby.