Fix Ruby VM interruptions causing `HttpTransport#export` failures
It took me a bit to wrap my head around this one, but
`rb_thread_call_without_gvl2` can return with an "interrupt".

This interrupt may mean that the VM has some work to do (and wants us
to call something like `rb_thread_check_ints()`), BUT that SHOULD NOT be
a reason for the export to fail, as this is expected to happen from
time to time.

I initially missed this subtlety, and it manifested itself as flaky
tests in CI (this seems to happen more often in Ruby 2.3 and below).
Tests would randomly fail with the
`Interrupted before call_exporter_without_gvl ran` message we
previously had.

After some experimentation, I found I could reproduce on any Ruby
version, if I just did enough calls to export:

```ruby
    context 'when server returns a 5xx failure' do
      let(:server_proc) { proc { |_req, res| res.status = 503 } }

      it 'logs an error' do
        expect(Datadog.logger).to receive(:error).with(/unexpected HTTP 503/)

        10_000.times do |i|
          print '@' if i % 100 == 0
          http_transport.export(flush)
        end
      end
    end
```

The fix here is to separate two different cases:
* There's an interrupt and the VM wants to process it. This shouldn't
  fail the export -- we just let the VM do whatever and retry our
  call.

* There's an exception to be raised. This will fail the export.