Handle VM interruptions during `StackRecorder` serialization instead of failing
After fixing the VM interruptions issue for `HttpTransport` in
4cd898e829d34f994973452e4503a88d37b4f99c (#1923), I suspected
that the same issue could happen to the `StackRecorder`.

It took a few attempts to reproduce this issue, but the
easiest way seems to be to "send ctrl+c to itself" by adding

```c
  #include <signal.h>

  kill(getpid(), SIGINT);
```

before the call to `rb_thread_call_without_gvl2`.

This made the VM flag an interrupt, which as long as no
exception was raised (and RSpec has its own interrupt handler)
simulated the issue.

TL;DR `rb_thread_call_without_gvl2` can return BEFORE running our
code when the VM has a pending interruption, so we need to let the VM
do its work, but there's no reason to fail the action (in this case,
fail to serialize) if all the VM wants is to run some callback
and then we can get on with our work.

So instead we retry running our code until it either runs or the VM
actually raises an exception.

The approach here is simpler than in `HttpTransport` because the
heap allocations happen only inside `call_serialize_without_gvl`;
in `HttpTransport` there are heap allocations done before calling
`rb_thread_call_without_gvl2` and thus a more complex approach is
needed.