Retrofit legacy profiling transport implementation as a backup to `HttpTransport` class
Everything in `lib/datadog/profiling/transport/` already existed prior
to the migration to the libddprof-based `HttpTransport` class
(#1923).

Temporarily, and just in case there are unknown-unknowns with
`HttpTransport`, I've re-imported a few classes, and modified them
so they will be able to be used **instead of**
`Datadog::Profiling::HttpTransport`.

This is only temporary, and will be deleted/reverted later.