Fix docs link not working
The old `http://gems.datadoghq.com/trace/docs/` link does not seem
to load for me on firefox nor chrome. I can see it with curl though...

I suspect this may be related to modern browsers trying to get to
an https version of the site, which doesn't work.

Instead, I've replaced the link with one that works over https and
seems to be working fine. In the future if/when we fix https support
for `gems.datadoghq.com` directly we can go back to the old url :)

Fixes #2057