Avoid Ruby 2.3 crash with `Floating point exception`
On purpose, this PR did not yet add anything to the
`struct per_thread_context`. This means that its `sizeof(...)` is 0,
and for all Rubies OTHER than 2.3 it's fine to do
`ruby_xcalloc(1, sizeof(struct per_thread_context))` even if the
struct is empty, but 2.3 doesn't like this and crashes with a
`Floating point exception`.

So for now let's just add a single dummy member to the struct to
unbreak CI, which will be replaced by the actual context we need in
a follow-up PR.