[PROF-5747] Fix broken libddprof linking on Heroku and AWS Elastic Beanstalk
As reported in #2067, in these environments ddtrace (and libddprof)
are moved after installation, which broke linking from the
profiling native extension to libddprof.

As a fix/"workaround", we additionally add the relative path
between both gems while linking; see the comments on the
`.libddprof_folder_relative_to_native_lib_folder` helper
for more details and how this works.

Note that key word above is **aditionally** -- e.g., we're
adding more paths in which to find libddprof, and keeping
the existing absolute path, so this should not impact
any setups where things were already working fine.

Big and special thanks to @sanchda for brainstorming with me
on this issue.

Fixes #2067