Add thanks to @miketheman for gifting us access to the `datadog` gem
We've long been in talks to adopt the `datadog` gem name for future
projects, but an issue was that the name was already taken on
rubygems.org

@miketheman was awesome in allowing us to get that name, and all
he asked in return was that his contribution not be forgotten, which
is extremely fair :)

So, let's thank him! And good luck for current/future endeavors!

P.s.: In the future, if/when we create a repository for the
`datadog` gem, let's remember to keep these thanks there as well!