Pin benchmark-memory to version usable in older Rubies
In all existing lockfiles used for CI testing, bundler/rubygems
resolved `benchmark-memory` to 0.1.2 which was version compatible
with all Rubies we use, as well as matching the
`memory_profiler` `~> 0.9` restriction we have.

But in some runs
<https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/6506/workflows/f1d66d70-dd6d-4940-81c9-21b2c1f683cc/jobs/240816>
I now saw the wrong version being resolved, so I've added a stronger
restriction to get bundler/rubygems to pick the version it should.