Add google-protobuf version restrictions to CI rack app
We had these restrictions for other apps, but not the rack app.

Instead, the rack app was being made to work using
`BUNDLE_FORCE_RUBY_PLATFORM`, but that was removed in the previous
commit.

This is getting to be an awful mess, but profiling will soon
not require protobuf so I don't think it's work investing in
improving for now; we just need it to hold for a little longer and
then we can remove it.