Add workaround for missing suspended events
I still need to investigate this further, but I believe this is a bug
in the GVL instrumentation API -- we observe threads being resumed
without the previous thread being suspended, which makes no sense.

Since the raw file output is timestamp-ordered, we can be sure
that if we see a resume event and the previous thread has not had
a suspended event, that it did suspend. So we inject a synthetic
event to make it easier to analyze the output.

I'll try to open an upstream issue with this discovery.