Inline list of classes that we monkey patch
The previous code was a bit hard to follow, since it had the whole
logic to pick the singleton_class vs class, and the whole thing
about accessing the toplevel binding.

In fact, all said and done, the affected classes were the ones
you can now see here. I've validated that the set of classes chosen
is correct and exactly matches the previous logic for Ruby 1.9 to 3.1.

In particular, I'm not sure why we had that whole thing about
the TOPLEVEL_BINDING.receiver and whatnot just to get the `Object`
class. Perhaps we hadn't noticed at the time that in the end
we really did get the `Object` class?