Simplify profiling `Forking` monkey patch, making it less generic
The profiling `Forking` monkey patch only has a single customer:
the `Datadog::Profiling::Tasks::Setup` class, which uses it to
add an `at_fork(:child)` that takes care of restarting the profiler
and fixing up the thread ids after a fork happened.

Since the `Forking` monkey patch affects core Ruby classes, I decided
to simplify it, as we're not using the added complexity, and I prefer
it being as simple as possible.

My additional motivation is that I'm working on support for monkey
patching `Process.daemon`, and this simplifies the amount of code
and tests needed to implement the needed functionality.

(We can always undo this if later we find a need for the other
callbacks, but in the ~2 years of life of this monkey patch, we
haven't needed the extra functionality, so I'll take that as a signal).