Ensure all tags are strings when reporting profile
A customer reported having their profiles fail with the following
issue:

> Unable to report profile. Cause: wrong argument type nil (expected String) Location: /var/cache/bundle/ruby/3.0.0/gems/ddtrace-1.2.0/lib/datadog/profiling/http_transport.rb:115:in _native_do_export'

Unfortunately there's a bunch of data passed in to that function so
just from that information it's not possible to be 100% sure which one
was `nil`, but while doing a second pass at the code I noticed one way
to trigger this would be to have tags that are `nil`.

Thus here's a quick fix to ensure that the tags passed on to
profiling are always strings.

Issue #2151