Add `samples_from_pprof` to profiling spec helper
There's quite a few specs where we need to take a pprof output and
assert on its contents, which are in a really awkward format for
most assertions we want.

So rather than repeating the conversion code everywhere, I've added
a nice helper that can be used accross multiple specs (and that
is quite useful when debugging stuff too).