[PROF-4902] Include 'thread id' label when sampling threads
In a previous PR, we added a `per_thread_context` structure that can
store information for each thread being sampled.

The first bit of information being stored there is the `thread id`,
which we will also use to tag all samples with the thread they came
from.

This behavior is the same as currently implemented by the
`OldStack` sampler.