[PROF-4902] Gather thread name during sampling
I've wanted to support this extra tiny feature for quite a while now:
gathering the thread names, not just their IDs.

We already have this feature for the Java and Python profilers, and I
like it quite a lot.

(Note: Since this feature is only being added on the profiling native
extension, it won't show up for customers for now since we haven't
wired it up for use yet.)