[PROF-4902] Track passage of wall-clock time between samples
One of the main responsibilities of the `CpuAndWallTime` collector
is tracking the amount of wall-clock time (aka real world time) that
elapses between samples.

It works by tracking the timestamp of when a thread was sampled
(in `wall_time_at_previous_sample_ns` inside the `per_thread_context`),
and then tagging samples with the elapsed time between the current and
the previous sample. (The first time a thread is sampled, the elapsed
time is set to 0).

This allows users to tell what their threads are spending time, even
when they're not actually spending CPU (e.g. when they're waiting on
a database query).

Tracking of cpu-time is still missing, and will be added in a separate
commit/PR.