Fix "function might be candidate for attribute noreturn" warning
The `raise_unexpected_type` function always raises an exception.

When compiling on Linux, I was getting this very helpful warning:
```
../../../../ext/ddtrace_profiling_native_extension/ruby_helpers.c: In function 'raise_unexpected_type':
../../../../ext/ddtrace_profiling_native_extension/ruby_helpers.c:3:6: warning: function might be candidate for attribute 'noreturn' [-Wsuggest-attribute=noreturn]
    3 | void raise_unexpected_type(
      |      ^~~~~~~~~~~~~~~~~~~~~
```

So to fix this warning, and make it clear to the compiler that the
function never returns, I've wrapped it with the `NORETURN` macro
that Ruby defines in the VM headers.