Tweak docker-compose setup for arm64 machines
I've been pairing with @TonyCTHsu on getting our docker-compose
to a state where it can run and pass the entire test suite when
run on an arm64 machine.

The following configuration is an improvement on the status quo
(as far as we know, only mongodb seems to be failing on arm64
at the moment), and still happily runs on x86-64, as well
still being green in CI.

Note for anyone running stuff on arm64: docker-compose/docker
only consider the `platform` setting when first downloading
an image, so you'll probably need to delete the existing
linux/x86_64 images for docker to realize it can download arm64
images instead.