[PROF-4902] Track passage of CPU-time between samples
In #2163 we added tracking of wall-clock time, and this commit
adds CPU-time, which is needed to produce CPU profiles aka to
figure out how much time different parts of the code spent executing
on CPU.

Like for the old profiler, we only support getting CPU-time on Linux.
I've added a fallback code path so that we can still easily develop
and test the profiler on macOS (even though we don't officially
support its use).

The way this works is similar to wall-clock time: on every sample
we get the current value from the OS, and we calculate the
elapsed CPU time from the previous sample
(which is stored in `cpu_time_at_previous_sample_ns`).
(The first time a thread is sampled, the elapsed time is set to 0).