Extract `StackRecorder` instance native state to struct
Previously, the only native state for the `StackRecorder` was the
libdatadog profile, and thus we directly stored a pointer to it
using the Ruby `TypedData` mechanism.

In preparation to store state beyond just a single libdatadog
profile, I've introduced a new struct and `StackRecorder` now wraps
the entire struct. This does not change any behavior, but opens
the door for us to introduce more state.