Add concurrency-safe mechanism to StackRecorder
The `StackRecorder` state is expected to be accessed in parallel by
two different threads (the sampler thread and the serializer thread).

To make such access correct even when it happens in parallel, this
commit adds a mutex-based mechanism to protect the `StackRecorder`
state.

See the added documentation for implementation details.