Replace `Exporter#empty?` with `Exporter#can_flush?`
The `#empty?` method was used by the `Scheduler` to determine if there
was data to be flushed.

With the move to libdatadog to keep the data, it's somewhat awkward to
implement an `#empty?` method, since that would entail looking at
internal libdatadog state or tracking such a thing ourselves.

Instead, let's replace it with a time-based `#can_flush?` that uses
the same `duration_below_threshold?` mechanism we had to decide if
there was enough data on a flush. The resulting outcome is similar,
but needs no extra tracking outside of the `Exporter` class.