Prepare `CpuAndWallTime` collector to be called by the upcoming worker class
In my original plans, I was thinking of having the time-based
trigger sampling behavior live in the `CpuAndWallTime` collector, but
as the collector has quite a bit of other responsibilities, I decided
to instead have that behavior as a separate class, called
`CpuAndWallTimeWorker`.

To be able to trigger sampling from that class, I've exposed that
function to other components in the project, and I've also added a
helper function that can be used in its constructor to ensure it got
a correctly-built `CpuAndWallTime` collector (the enforce function).