RFC: Refactor how native extension testing methods are exposed
This change follows from [this discussion
thread](https://github.com/DataDog/dd-trace-rb/pull/2181#discussion_r929383004)
on improving how testing-only methods are exposed from the native
extension.

Here's what I came up with:

* All methods exposed only for testing are scoped inside `Testing`
  modules; e.g. testing methods for `StackRecorder` are inside
  `StackRecorder::Testing`

* Removed the "middlemen" methods on the classes themselves. They
  existed only for documentation of the existence and objective of those
  methods, but now that they are scoped under `Testing` the
  "middlemen" methods seem less interesting to have.

* Changed any specs using those methods to directly call the native
  methods in the `Testing` modules.

* Still kept the `_native_` prefix for all native methods that get
  exposed to Ruby. I'm open to discussing this, but I really like being
  extremely explicit on when the Ruby/C boundary is being crossed.

This doesn't change anything for production code, it only shuffles test
code around. I'm opening it as an RFC to make it clear this is an
option that seems nice but I'm open to changing it further.