Minor improvement to timeout documentation
As raised in #818, the default `timeout` example is rather far away
from the actual default being used, which may lead customers that
copy-paste the example to have a bad experience.

I also discovered we had an incorrect constant with the default
timeout which was not even being used anywhere (you can grep for it
 to confirm!).

Thanks @benjaminwood for the awesome suggestion!