[PROF-5859] Introduce `CpuAndWallTimeWorker` to trigger profiling samples
The `CpuAndWallTimeWorker` component takes care of automatically
triggering the `CpuAndWallTime` collector.

In a single sentence, this component a) runs in a background thread,
b) periodically uses a unix signal (SIGPROF) to interrupt a running
Ruby thread, and c) gets the interrupted Ruby thread to call the
`CpuAndWallTime` collector to take a sample.

And thus we achieve periodic sampling of Ruby threads. The
`CpuAndWallTimeWorker` was the last big missing piece for the new
Ruby profiler, and we can now turn it on and actually get samples.

The inner workings of the `CpuAndWallTimeWorker` are non-trivial,
and hopefully the design description included, as well as the specs
will help in documenting what it does and how it does it.

Finally, I would consider this component as not being complete yet,
and I'll call the current state "alpha".

In particular, it's missing at least:

* Handling of forks (there's some work, but I'm not convinced it's
  enough)
* Validation / specs that show it's Ractor-safe
* Dynamic sampling rate/pacing mechanism
* Sending signals to specific threads, not to the whole process,
  to reduce bias and improve CPU profiling

I have notes for all of these things, and will continue working on
them until we're confident on the implementation and on our
validation, and can start asking customers to test it out.