Tweak the upgrade guide banner in README/Getting Started to mention 1.x instead of 1.0.0
**What does this PR do?**:

Tweak the language of the note we added at the top of the README.md and
GettingStarted.md documents to refer to the 1.x series, and not just
to 1.0.0.

**Motivation**:

It seemed a bit weird/confusing to me that the wording was presenting
1.0.0 as if it was the latest release, whereas customers should be
on 1.3.0 (or whatever comes after).

**How to test the change?**:

This does not affect code, so looking at the diff and the GitHub
markdown preview is enough.