Fix outdated references to `use` instead of `instrument` in docs
While reviewing #2217 where we fixed the sinatra documentation not to
mention the outdated `use`, I decided to make a quick check if we had
any more such leftovers, and it turns out we did.

This PR fixes the two leftover mentions of `use` that I found that
should be replaced by `instrument`. I did not change the sinatra docs
to avoid a conflict with #2217.