[SCP-184] Fix incompatible library check when there's no libruby.so
The logic in `ddtrace_profiling_loader.c` follows Ruby's `dln.c`;
see the big comment at the top of the former for why it needs to exist.

The `incompatible_library` sanity check exists to try to catch
situations where the library being loaded is linked to a different
`libruby.so` (or equivalent in other OSs) than the one that is doing
the loading.

BUT when we implemented `incompatible_library` we missed an check
that upstream Ruby does.

From https://github.com/ruby/ruby/blob/v3_1_2/dln.c#L295:

```c
static bool
dln_incompatible_library_p(void *handle)
{
    void *ex = dlsym(handle, EXTERNAL_PREFIX"ruby_xmalloc");
    void *const fp = (void *)ruby_xmalloc;
    return ex && ex != fp;
}
```

Ruby actually checks that the symbol found is **not NULL**
(e.g. with the `ex && ` check). This is important, because there's
a valid situation in which the symbol is expected to be NULL:
when Ruby is built with `--disable-shared` at compilation time.

What `--disable-shared` does is that it builds the Ruby VM logic
inside the `ruby` executable itself INSTEAD OF splitting the logic
between the `ruby` executable and the `libruby.so` shared library.

To properly support this configuration, we should, like upstream
Ruby does, consider a library compatible if the symbol lookup
returns `NULL`.

Extra note -- here's a few ways of telling if there's a `libruby`:

* `RbConfig::CONFIG['configure_args']` will contain `--disable-shared`
* `RbConfig::CONFIG['LIBRUBY']` will be `libruby-static.a` instead
   of, for instance `libruby.so.3.1.1`
* Ruby will not be linked to libruby:
  ```
  # with libruby
  $ ldd `which ruby` | grep libruby
    libruby.so.3.1 => /usr/local/lib/libruby.so.3.1 (0x00007fb37e260000)

  # without libruby
  $ ldd `which ruby` | grep libruby
  (no output)
  ```