Update `HttpTransport` integration tests from intake v1.3 to v2.4 format
Libdatadog 0.9 reports data using the intake v2.4 format, and we
were previously using v1.3.

Thus, the integration specs need to be updated to match the new format.

One additional change also included here is testing that the file
payloads are compressed using LZ4 (which is also a new libdatadog 0.9
feature).