Support placeholder Garbage Collection frames in Collectors::Stack
As part of showing CPU and wall-clock time spent in garbage collection,
let's extend the `Collectors::Stack` to be able to add a placeholder
"Garbage Collection" frame when sampling the stack.

This is exposed as a new `sample_thread_in_gc`, which will be used
by the `Collectors::CpuAndWallTime` in a later commit.