[PROF-6260] Add support for tracking time spent in GC to CpuAndWallTime collector, second try
**What does this PR do?**:

This PR builds on top of #2304 to deliver the new "showing time spent in GC in
the flamegraph feature". It does not yet offer the full solution: it is still
missing the part where we integrate with the Ruby VM to automatically track the
in-gc/not-in-gc state for threads. This final bit will come in a follow-up PR.

Specifically, this PR modifies the cpu and wall-time tracking in the
`CpuAndWallTime` collector to introduce a new implicit state that a thread can
be in: doing garbage collection.

This works by introducing three new functions:

* `cpu_and_wall_time_collector_on_gc_start`
* `cpu_and_wall_time_collector_on_gc_finish`
* `cpu_and_wall_time_collector_sample_after_gc`

When `cpu_and_wall_time_collector_on_gc_start` gets called, the current cpu and
wall-time get recorded to the thread context: `cpu_time_at_gc_start_ns` and
`wall_time_at_gc_start_ns`.

While these fields are set, regular samples (if any) do not account for any time
that passes after these two timestamps.

(Regular samples can still account for the time between the previous sample and
the start of GC.)

When `cpu_and_wall_time_collector_on_gc_finish` gets called, the current cpu and
wall-time again get recorded to the thread context: `cpu_time_at_gc_finish_ns`
and `wall_time_at_gc_finish_ns`.

Finally, when `cpu_and_wall_time_collector_sample_after_gc` gets called, the
following happens:

1. A sample gets taken, using the special `SAMPLE_IN_GC` sample type, which
produces a stack with a placeholder `Garbage Collection` frame as the latest
frame. This sample gets assigned the cpu-time and wall-time period that was
recorded between calls to `on_gc_start` and `on_gc_finish`.

2. The thread is no longer marked as being in gc (all gc tracking fields get
reset back to `INVALID_TIME`).

3. The `cpu_time_at_previous_sample_ns` and `wall_time_at_previous_sample_ns`
get updated with the elapsed time in GC, so that all time is accounted for --
e.g. the next sample will not get "blamed" by time spent in GC.

Looking at the history and at the name of this PR, you may notice that this
is the "second try". The first try was never merged in #2308, and it
unfortunately suffered from a flaw when used in the context of Ruby's
garbage collection hooks -- it could trigger memory allocation in
rare occasions (usually exceptions), which is actually not allowed to happen
during Ruby's garbage collection start/finish hooks.

So I needed to rewrite the approach to make sure that `on_gc_start` and
`on_gc_finish` can never allocate memory, and to move the sample generation to
`sample_after_gc`, which the `Collectors::CpuAndWallTimeWorker` will make sure
to call as soon as "safe" after GC ends.

**Motivation**:

This work will enable us to show the time spent in garbage collection in the
flamegraph.

This time would previously be invisible, and accounted for as time for the
unlucky Ruby method that was running when GC happened.

**Additional Notes**:

As discussed above, this PR is only a piece of the puzzle, and is still
missing changes to the `CpuAndWallTimeWorker´ to integrate with the Ruby VM.

**How to test the change?**:

The change includes code coverage.